+++ tests/src/main/java/org/nervos/appchain/tests/TokenTransactionExample.java
@@ -24,7 +24,7 @@ import org.nervos.appchain.protocol.core.methods.response.TransactionReceipt;
 
 
 public class TokenTransactionExample {
-    private static int chainId;
+    private static BigInteger chainId;
     private static int version;
     private static String privateKey;
     private static String fromAddress;
